backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
PyJWT==2.10.1
torch==2.5.1
pytest==8.3.4
httpx==0.28.1
